@@ -2,7 +2,7 @@
     "name": "laravel-goto-lang",
     "displayName": "Laravel Goto Lang",
     "description": "Quick jump to lang",
-    "version": "0.3.1",
+    "version": "0.3.2",
     "publisher": "ctf0",
     "engines": {
         "vscode": "^1.52.0"
@@ -79,29 +79,28 @@
             }
         }
     },
-    "main": "./out/src/extension",
+    "main": "./out/main",
     "scripts": {
-        "vscode:prepublish": "yarn run compile",
-        "compile": "tsc -p ./",
-        "watch": "tsc -watch -p ./",
-        "pretest": "yarn run compile",
-        "test": "node ./out/test/runTest.js"
+        "vscode:prepublish": "npm run esbuild-base -- --minify",
+        "esbuild-base": "esbuild ./src/extension.ts --bundle --outfile=out/main.js --external:vscode --format=cjs --platform=node",
+        "esbuild": "npm run esbuild-base -- --sourcemap",
+        "esbuild-watch": "npm run esbuild-base -- --sourcemap --watch",
+        "watch": "npm run vscode:prepublish",
+        "test-compile": "tsc -p ./"
     },
     "devDependencies": {
-        "@types/glob": "^7.2.0",
-        "@types/mocha": "^9.1.0",
-        "@types/node": "^17.0.23",
+        "@types/glob": "^8.0.0",
+        "@types/node": "^18.11.9",
         "@types/vscode": "^1.52.0",
-        "mocha": "^9.2.2",
+        "esbuild": "^0.15.13",
         "tslint": "^5.20.1",
-        "typescript": "^4.6.3",
-        "vscode-test": "^1.5.0"
+        "typescript": "^4.8.4"
     },
     "dependencies": {
         "await-exec": "^0.1.2",
-        "escape-string-regexp": "^4.0.0",
-        "fast-glob": "^3.2.11",
+        "escape-string-regexp": "^5.0.0",
+        "fast-glob": "^3.2.12",
         "lodash": "^4.17.21",
-        "pascalcase": "^1.0.0"
+        "pascalcase": "^2.0.0"
     }
 }
